SplitRedisStorage in diferent file
