+++ deprecated/erniboot.sh
